pyproject.toml
[project]
name = "func-blob-to-index"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "azure-functions (>=1.23.0,<2.0.0)",
    "azurefunctions-extensions-bindings-blob (>=1.0.0,<2.0.0)",
    "openai (>=1.97.0,<2.0.0)",
    "azure-search-documents (>=11.5.3,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
